adding melting temp observable (#4)
This change adds the melting temp observable class for computing melting temp using umbrella sampling. Primarily this change is restricted to the melting_temp module and the tests for that (along with some supporting test data), but it also adds an oxdna template (tm-6bp-2op) for our working example.

In addition, we add an example that uses this class for optimization (tm_optimization.py), along with changes to the difftre objective to allow different observable types, supporting optimization of melting temp. more unit conversion functions for temperature function for reading oxdna energy file.